Fix the location that selecting a mark uses
I think this subtly regressed in

`Terminal::SelectNewRegion` is the only thing that uses the return value from
`Terminal::_ScrollToPoints`. Before that PR, `_ScrollToPoints` was just a part
of `SelectNewRegion`, and it moved the start & end coords by the
`_VisibleStartIndex`, not the `_scrollOffset`.

Kinda weird there weren't any _other_ tests for `SelectNewRegion`?

I also caught a second bug while I was here - If you had a line with an exact
wrap, and tried to select that like with selectOutput, we'd explode.

Closes #17131